# [TRA tests] Speed up eval fixed cands test