**Patch description**
Decrease number of epochs that the eval_fixed_cands test runs for. As far as I can tell this doesn't affect whether or not the test passes usually. For the three models that run this test, the speedups are as follows:

**Before**
```
[ Test test_eval_fixed completed in 10.13738s ]
.
[ Test test_eval_fixed completed in 18.37439s ]
..
[ Test test_eval_fixed completed in 14.78047s ]
```

**After**
```
[ Test test_eval_fixed completed in 2.79047s ]
........
[ Test test_eval_fixed completed in 6.67095s ]
.........
[ Test test_eval_fixed completed in 5.17016s ]
```

I also added a decorator to be able to easily track this when you are debugging. Idk if generally useful, up to @stephenroller ...